small modification in files names and imports

+++ projects/back/src/main/java/com/exerc/mybackend/repositories/HousingLocationRepository.java
@@ -4,5 +4,5 @@ import com.exerc.mybackend.model.HousingLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.exerc.back.entities.HousingLocation;
 
-public interface HousingLocationRepository extends JpaRepository<HousingLocation, Long> {
+public interface HousingLocationRepository extends JpaRepository<HousingLocation, Integer> {
 }
